refactor(app): split the tool dispatch leaf out of the catalog
check-import-cycles caught a real cycle I had not run locally: catalog.ts
value-imported run-code.ts for the run_code entry, and run-code.ts
value-imported catalog.ts for cellDoor.

The fix is not a shim. Two different things were living in one file: what a
tool entry IS and how a name runs one (dispatch.ts, which knows no tool), and
WHICH tools this app has (catalog.ts, which knows every tool). The cell door
moved to run-code.ts, where its only consumer already was and where the
machine-tool question belongs.
